requirements.txt
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-prometheus==2.3.1
